Additional check if the context allows trusted computers
It is also required to check if the context allows to bypass the 2FA check if the device is trusted.
It is to be used in scenarios when the context resolves the ability to "trust" the device in runtime (say - depending on the client's network). In those cases even if the user has the trusted device - it may be not trusted under this very circumstances.