[RFC] TOTP doesn't match TOTPInterface
First up, thanks for maintaining this!

I noticed after updating to v11 (which included replaced TOTP window with
leeway parameter) that the following code no longer satisfies my PHPStan
static analysis.

```
use OTPHP/TOTP;
$totp = TOTP::create($Secret);
$Verify = $totp->verify($In['otp'], leeway: 20);
```

The problem is that TOTP::create() returns a TOTPInterface,
which extends OTPInterface. OTPInterface->verify() has a window parameter,
and not a leeway parameter.

I've done a minimally invasive fix here (only for TOTP), but looking
at the code, the interfaces don't match the functions in other areas
as well.

My honest opinion is that most of the interfaces are a liability here.
We already have an abstract OTP class implementing the common methods
and we can define abstract methods there to force extending classes to
implement methods required for the factory.

If you're open to it, I would be happy to put together a different
pull request to demonstrate.